Fix all Vulkan-backend test failures (196 → 0): dtype parity, QR, MaskRCNN, stress conditioning
Brings the Vulkan compute backend to cross-backend parity with CPU/CUDA/ROCm/
OneAPI and fixes several real correctness bugs that were masked by a box_iou bug.

Vulkan dtype/numerical coverage:
- Width-correct fill/creation/indexing/cast for Int8/16, UInt16/32/64,
  Complex64/128, FP8, quantized (new full_/strided_fill_/gather_/scatter_/
  cast_/arange_/linspace_ shaders).
- Integer-widen + complex real-decomposition matmul/bmm/dot; complex linalg.
- NaN-propagating max/min reductions; UInt64 reductions via f64.
- LayerNorm f64 mean/rstd outputs; logcumsumexp_f64 true double exp/log series.
- Allocator OOM fix: flush the deferred-free list on synchronize() and retry
  on allocation failure (cleared training-loop / large-model OOM).
- Optimizer input-order (RMSprop/Adagrad), NestedTensor offset segfault,
  EmbeddingBag last-bag, FFT/pooling f16 widen, FusedAttention GQA+causal compose.

QR (linalg_qr_panel{,_f64}.comp):
- Normalize the sub-diagonal Householder vector by (x[k]-alpha) so the stored
  reflectors match tau's v[k]=1 convention; previously Q@R != A.

Strided fill on sub-aligned views (strided_fill_i8.comp + vulkan_ops_misc.cpp):
- Bind the storage base and pass the view element offset to the shader; a
  1-byte-offset Int8 slice cannot be bound directly (SSBO offset alignment).

box_iou (src/ops/detection.cpp):
- .contiguous() the inputs; a non-contiguous column slice (proposals[:,1:5],
  row stride 5) was indexed as packed 4/row, corrupting the IoU matrix.

MaskRCNN (src/models/mask_rcnn.cpp):
- add_gt_proposals: append GT boxes to proposals so every object gets a
  positive ROI (otherwise a fresh RPN yields zero positives, mask_loss==0).
- Encode ROI-head box-regression targets as deltas relative to their proposals
  instead of comparing predicted deltas against raw GT coordinates.

Tests:
- Seed model init in MaskRCNNLossTest::SetUp for deterministic, reproducible
  weights across parameterized backend instances.
- DeepGraph_Residual: seed + reduce weight scale 0.1->0.02. At 0.1 the residual
  graph has gain ~1.9 and explodes to ~1e14 (chaotic), making cross-backend
  parity meaningless (Vulkan matched CPU exactly while CUDA drifted). 0.02 keeps
  activations bounded so the deep-graph parity check is well-conditioned.
- ConvNeXt depthwise: include Vulkan in the jitter-tolerance branch.

+++ src/backends/vulkan/vulkan_backend.hpp
@@ -306,7 +306,7 @@ public:
     // Kernel dispatch helpers — used by vulkan_kernel_registry.cpp
     auto dispatchBinaryOp(const std::string& op_name, const Tensor& a, const Tensor& b) -> Tensor;
     auto dispatchUnaryOp(const std::string& op_name, const Tensor& input) -> Tensor;
-    auto dispatchUnaryOpWithParam(const std::string& op_name, const Tensor& input, float param) -> Tensor;
+    auto dispatchUnaryOpWithParam(const std::string& op_name, const Tensor& input, double param) -> Tensor;
     // NanToNum and bitwise shift dispatchers (vulkan_ops_misc.cpp). Replace previous CPU fallbacks.
     auto dispatchNanToNum(const Tensor& input, float nan_val, float posinf_val, float neginf_val) -> Tensor;
     auto dispatchBitwiseBinaryOp(const std::string& shader_name, const Tensor& a, const Tensor& b) -> Tensor;
@@ -629,8 +629,8 @@ public:
     auto dispatchContiguous(const Tensor& input) -> Tensor;
 
     // Creation operations
-    auto dispatchArange(float start, float end, float step, DType dtype, const Device& device) -> Tensor;
-    auto dispatchLinspace(float start, float end, int64_t steps, DType dtype, const Device& device) -> Tensor;
+    auto dispatchArange(double start, double end, double step, DType dtype, const Device& device) -> Tensor;
+    auto dispatchLinspace(double start, double end, int64_t steps, DType dtype, const Device& device) -> Tensor;
     auto dispatchEye(int64_t n, int64_t m, DType dtype, const Device& device) -> Tensor;
 
     // Memory operations
@@ -749,7 +749,7 @@ public:
     auto dispatchMode(const Tensor& input, int64_t dim, bool keepdim) -> std::vector<Tensor>;
 
     // Misc operations (Phase 11.5)
-    auto dispatchStridedFill(Tensor& input, float value) -> void;
+    auto dispatchStridedFill(Tensor& input, double value) -> void;
     auto dispatchToMemoryFormat(const Tensor& input, int format) -> Tensor;
     auto dispatchHasInfNan(const Tensor& input) -> Tensor;
     auto dispatchDepthwiseConv2d(const Tensor& input, const Tensor& weight,
@@ -786,7 +786,7 @@ public:
     // Tensor manipulation operations (Repeat/Masked already declared above)
     auto dispatchExpand(const Tensor& input, const std::vector<int64_t>& shape) -> Tensor;
     auto dispatchCat(const std::vector<Tensor>& inputs, int64_t dim) -> Tensor;
-    auto dispatchClamp(const Tensor& input, float min_value, float max_value) -> Tensor;
+    auto dispatchClamp(const Tensor& input, double min_value, double max_value) -> Tensor;
 
     // Interpolation operation
     auto dispatchInterpolate(const Tensor& input, const OpAttributes& attrs) -> Tensor;
